argmin_math

Trait ArgminScaledAdd

pub trait ArgminScaledAdd<T, U, V> {
    // Required method
    fn scaled_add(&self, factor: &U, vec: &T) -> V;
}

Add a T scaled by an U to self

Required Methods

fn scaled_add(&self, factor: &U, vec: &T) -> V

Add a T scaled by an U to self