Trait argmin_math::ArgminScaledAdd

pub trait ArgminScaledAdd<T, U, V> {
    fn scaled_add(&self, factor: &U, vec: &T) -> V;
}

Add a T scaled by an U to self

Required methods

fn scaled_add(&self, factor: &U, vec: &T) -> V

Add a T scaled by an U to self