argmin_math
Trait ArgminSignum

pub trait ArgminSignum {
    // Required method
    fn signum(self) -> Self;
}

Returns a number that represents the sign of self.

Required Methods

fn signum(self) -> Self

Returns a number that represents the sign of self.

Object Safety

This trait is not object safe.