argmin_math

Trait ArgminSignum

pub trait ArgminSignum {
    // Required method
    fn signum(self) -> Self;
}

Returns a number that represents the sign of self.

Required Methods

fn signum(self) -> Self

Returns a number that represents the sign of self.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.