Trait argmin_math::ArgminSignum

pub trait ArgminSignum {
    fn signum(self) -> Self;
}

Returns a number that represents the sign of self.

Required Methods

fn signum(self) -> Self

Returns a number that represents the sign of self.